+++ packages/api/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@plutomi/api",
+  "version": "0.1.0",
+  "description": "API Package for Plutomi.",
+  "main": "index.js",
+  "repository": "https://github.com/plutomi/plutomi",
+  "author": "Plutomi Inc.",
+  "license": "Apache-2.0",
+  "private": false,
+  "dependencies": {
+    "express": "^4.18.2",
+    "zod": "^3.21.4"
+  },
+  "scripts": {
+    "dev": "ts-node-dev --respawn --transpile-only index.ts"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.15.11",
+    "dotenv": "^16.0.3",
+    "ts-node-dev": "^2.0.0"
+  }
+}
