+++ packages/infra/env.ts
@@ -1,19 +1,19 @@
-import { z } from "zod";
+// import { z } from "zod";
 
-const schema = z.object({
-  DEPLOYMENT_ENVIRONMENT: z.enum(["prod", "stage", "dev"]),
-  DOMAIN: z.string().url(),
-});
+// const schema = z.object({
+//   DEPLOYMENT_ENVIRONMENT: z.enum(["prod", "stage", "dev"]),
+//   DOMAIN: z.string().url(),
+// });
 
-const parsed = schema.safeParse(process.env);
+// const parsed = schema.safeParse(process.env);
 
-if (!parsed.success) {
-  parsed.error.issues.forEach((issue) => {
-    console.error("\n❌ Invalid environment variable:");
-    console.error(issue);
-  });
+// if (!parsed.success) {
+//   parsed.error.issues.forEach((issue) => {
+//     console.error("\n❌ Invalid environment variable:");
+//     console.error(issue);
+//   });
 
-  process.exit(1);
-}
+//   process.exit(1);
+// }
 
-export const env = parsed.data;
+// export const env = parsed.data;
