--- packages/infra/index.ts
@@ -1,9 +1,0 @@
-import { env } from "./env";
-
-const main = () => {
-  console.log("Hello World!");
-
-  console.log("heres are my vars", env);
-};
-
-main();
