+++ plutomi.code-workspace
@@ -14,6 +14,6 @@
     { "name": "docs", "path": "packages/docs" }
   ],
   "settings": {
-    "cSpell.words": ["cacheable", "Plutomi", "tailwindcss"]
+    "cSpell.words": ["cacheable", "fargate", "Plutomi", "tailwindcss"]
   }
 }
